requirements.txt
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
scikit_learn==1.4.1.post1
